Add equipment category to API.
Also refactored stuff related to equipment API to separate module.

Refs #54
